@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AuthController } from './auth.controller';
+import { Test, TestingModule } from '@nestjs/testing'
+import { AuthController } from './auth.controller'
 
 describe('AuthController', () => {
-  let controller: AuthController;
+  let controller: AuthController
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [AuthController],
-    }).compile();
+      controllers: [AuthController]
+    }).compile()
 
-    controller = module.get<AuthController>(AuthController);
-  });
+    controller = module.get<AuthController>(AuthController)
+  })
 
   it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
+    expect(controller).toBeDefined()
+  })
+})
