@@ -1,7 +1,7 @@
 export class auth_dto {
-    email: string
-    password: string
-    cpassword: string
-    xaccesstoken: string
-    decodeid: string
+  email: string
+  password: string
+  cpassword: string
+  xaccesstoken: string
+  decodeid: string
 }
