@@ -1,8 +1,7 @@
-import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
-import { JwtModule } from '@nestjs/jwt';
-
+import { Module } from '@nestjs/common'
+import { AuthService } from './auth.service'
+import { AuthController } from './auth.controller'
+import { JwtModule } from '@nestjs/jwt'
 
 @Module({
   imports: [JwtModule.register({
@@ -13,4 +12,3 @@ import { JwtModule } from '@nestjs/jwt';
   controllers: [AuthController]
 })
 export class AuthModule { }
-
