@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { LevelController } from './level.controller';
+import { Test, TestingModule } from '@nestjs/testing'
+import { LevelController } from './level.controller'
 
 describe('LevelController', () => {
-  let controller: LevelController;
+  let controller: LevelController
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [LevelController],
-    }).compile();
+      controllers: [LevelController]
+    }).compile()
 
-    controller = module.get<LevelController>(LevelController);
-  });
+    controller = module.get<LevelController>(LevelController)
+  })
 
   it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
+    expect(controller).toBeDefined()
+  })
+})
