@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { LevelService } from './level.service';
+import { Test, TestingModule } from '@nestjs/testing'
+import { LevelService } from './level.service'
 
 describe('LevelService', () => {
-  let service: LevelService;
+  let service: LevelService
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [LevelService],
-    }).compile();
+      providers: [LevelService]
+    }).compile()
 
-    service = module.get<LevelService>(LevelService);
-  });
+    service = module.get<LevelService>(LevelService)
+  })
 
   it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
+    expect(service).toBeDefined()
+  })
+})
