@@ -1,19 +1,17 @@
-import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
-import * as cookieParser from 'cookie-parser';
-import { Post } from '@nestjs/common';
+import { NestFactory } from '@nestjs/core'
+import { AppModule } from './app.module'
+import * as cookieParser from 'cookie-parser'
+import { Post } from '@nestjs/common'
 // somewhere in your initialization file
 
-
-async function bootstrap() {
-
-  const app = await NestFactory.create(AppModule);
-  app.use(cookieParser());
+async function bootstrap () {
+  const app = await NestFactory.create(AppModule)
+  app.use(cookieParser())
   app.enableCors({
     origin: true,
     methods: 'GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS',
-    credentials: true,
-});
-  await app.listen(5000);
+    credentials: true
+  })
+  await app.listen(5000)
 }
-bootstrap();
+bootstrap()
