@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ModuleController } from './module.controller';
+import { Test, TestingModule } from '@nestjs/testing'
+import { ModuleController } from './module.controller'
 
 describe('ModuleController', () => {
-  let controller: ModuleController;
+  let controller: ModuleController
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [ModuleController],
-    }).compile();
+      controllers: [ModuleController]
+    }).compile()
 
-    controller = module.get<ModuleController>(ModuleController);
-  });
+    controller = module.get<ModuleController>(ModuleController)
+  })
 
   it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
+    expect(controller).toBeDefined()
+  })
+})
