@@ -1,4 +1,4 @@
 export class module_dto {
-    module: string
-    status: boolean
+  module: string
+  status: boolean
 }
