@@ -1,4 +1,4 @@
 import * as bcrypt from 'bcrypt'
-const saltOrRounds = 10;
-const password = 'random_password';
+const saltOrRounds = 10
+const password = 'random_password'
 // const hash = await bcrypt.hash(password, saltOrRounds);
