@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ParticipantsController } from './participants.controller';
+import { Test, TestingModule } from '@nestjs/testing'
+import { ParticipantsController } from './participants.controller'
 
 describe('ParticipantsController', () => {
-  let controller: ParticipantsController;
+  let controller: ParticipantsController
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [ParticipantsController],
-    }).compile();
+      controllers: [ParticipantsController]
+    }).compile()
 
-    controller = module.get<ParticipantsController>(ParticipantsController);
-  });
+    controller = module.get<ParticipantsController>(ParticipantsController)
+  })
 
   it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
+    expect(controller).toBeDefined()
+  })
+})
