@@ -1,8 +1,8 @@
-import { Module } from '@nestjs/common';
-import { ParticipantsController } from './participants.controller';
-import { ParticipantsService } from './participants.service';
+import { Module } from '@nestjs/common'
+import { ParticipantsController } from './participants.controller'
+import { ParticipantsService } from './participants.service'
 @Module({
-    controllers: [ParticipantsController],
-    providers: [ParticipantsService]
+  controllers: [ParticipantsController],
+  providers: [ParticipantsService]
 })
 export class ParticipantsModule { }
