@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ParticipantsService } from './participants.service';
+import { Test, TestingModule } from '@nestjs/testing'
+import { ParticipantsService } from './participants.service'
 
 describe('ParticipantsService', () => {
-  let service: ParticipantsService;
+  let service: ParticipantsService
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [ParticipantsService],
-    }).compile();
+      providers: [ParticipantsService]
+    }).compile()
 
-    service = module.get<ParticipantsService>(ParticipantsService);
-  });
+    service = module.get<ParticipantsService>(ParticipantsService)
+  })
 
   it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
+    expect(service).toBeDefined()
+  })
+})
