@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { QuestionOptionsController } from './question_options.controller';
+import { Test, TestingModule } from '@nestjs/testing'
+import { QuestionOptionsController } from './question_options.controller'
 
 describe('QuestionOptionsController', () => {
-  let controller: QuestionOptionsController;
+  let controller: QuestionOptionsController
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [QuestionOptionsController],
-    }).compile();
+      controllers: [QuestionOptionsController]
+    }).compile()
 
-    controller = module.get<QuestionOptionsController>(QuestionOptionsController);
-  });
+    controller = module.get<QuestionOptionsController>(QuestionOptionsController)
+  })
 
   it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
+    expect(controller).toBeDefined()
+  })
+})
