@@ -1,8 +1,8 @@
 export class question_dto {
-    option_type: string
-    option:string
-    question_id:string
-    correct:Boolean
-    image:string
-    status:Boolean
+  option_type: string
+  option: string
+  question_id: string
+  correct: Boolean
+  image: string
+  status: Boolean
 }
