@@ -1,6 +1,6 @@
-import { Module } from '@nestjs/common';
-import { QuestionOptionsController } from './question_options.controller';
-import { QuestionOptionsService } from './question_options.service';
+import { Module } from '@nestjs/common'
+import { QuestionOptionsController } from './question_options.controller'
+import { QuestionOptionsService } from './question_options.service'
 
 @Module({
   controllers: [QuestionOptionsController],
