@@ -1,18 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { QuestionsController } from './questions.controller';
+import { Test, TestingModule } from '@nestjs/testing'
+import { QuestionsController } from './questions.controller'
 
 describe('QuestionsController', () => {
-  let controller: QuestionsController;
+  let controller: QuestionsController
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [QuestionsController],
-    }).compile();
+      controllers: [QuestionsController]
+    }).compile()
 
-    controller = module.get<QuestionsController>(QuestionsController);
-  });
+    controller = module.get<QuestionsController>(QuestionsController)
+  })
 
   it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
+    expect(controller).toBeDefined()
+  })
+})
