@@ -1,14 +1,12 @@
 class QuestionDTO {
-    id: Int32Array;
-    question: string
-    question_type: string
-    question_time: Date
-    images:string
-    createdAt: Date
-    updatedAt: Date
-    status: boolean
+  id: Int32Array
+  question: string
+  question_type: string
+  question_time: Date
+  images: string
+  createdAt: Date
+  updatedAt: Date
+  status: boolean
+}
 
-  }
-  
-  export { QuestionDTO };
-  
+export { QuestionDTO }
